mapping API version 1.1
disabled ability to change routing info by clicking on the map as that would be unintended behavior of a manager tracking robot movement. Hard set MLK library as the origin of the directions for the imaginary store location. Next need to be able to receive valid address from the server to feed to API program to make directions hard set and uninfluenced by the user on the map.